# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/collects": {
            "post": {
                "tags": [
                    "Collect"
                ],
                "summary": "Add product to collection",
                "description": "Add product to collection",
                "operationId": "addProductToCollection",
                "requestBody": {
                    "description": "Product object that needs to be added to collection",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id, collection_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "uuid",
                                        "example": "c0a80121-7ac0-18d3-8e47-000000000000"
                                    },
                                    "collection_id": {
                                        "type": "uuid",
                                        "example": "c0a80121-7ac0-18d3-8e47-000000000000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CollectResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Collection not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The product id field is required.",
                                                "The collection id field is required."
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "description": "Returns list of products",
                "operationId": "getProducts",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create product",
                "description": "Create product",
                "operationId": "createProduct",
                "requestBody": {
                    "description": "Product object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The name field is required."
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product by id",
                "description": "Returns product by id",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "schema": {
                            "type": "uuid",
                            "example": "ea9f1b1a-1b1a-11eb-adc1-0242ac120002"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update product",
                "description": "Update product",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "schema": {
                            "type": "uuid",
                            "example": "ea9f1b1a-1b1a-11eb-adc1-0242ac120002"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Product object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The name field is required."
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductRequest": {
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "description": "Product title",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Product title"
                    },
                    "status": {
                        "description": "Product status",
                        "type": "string",
                        "default": "active",
                        "enum": [
                            "active",
                            "draft",
                            "archived"
                        ],
                        "example": "active"
                    },
                    "options": {
                        "description": "Product options",
                        "type": "array",
                        "items": {
                            "properties": {
                                "name": {
                                    "description": "Option name",
                                    "type": "string",
                                    "maxLength": 255,
                                    "example": "Color"
                                },
                                "values": {
                                    "description": "Option values",
                                    "type": "array",
                                    "items": {
                                        "description": "Option value",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Red"
                                    },
                                    "example": [
                                        "Red",
                                        "Blue"
                                    ]
                                }
                            },
                            "type": "object"
                        },
                        "maxItems": 3,
                        "example": [
                            {
                                "name": "Color",
                                "values": [
                                    "Red",
                                    "Blue"
                                ]
                            }
                        ]
                    },
                    "variants": {
                        "description": "Product variants",
                        "type": "array",
                        "items": {
                            "properties": {
                                "option1": {
                                    "description": "Option 1",
                                    "type": "string",
                                    "maxLength": 255,
                                    "example": "Red"
                                },
                                "option2": {
                                    "description": "Option 2",
                                    "type": "string",
                                    "maxLength": 255,
                                    "example": "Small"
                                },
                                "option3": {
                                    "description": "Option 3",
                                    "type": "string",
                                    "maxLength": 255,
                                    "example": "Round"
                                }
                            },
                            "type": "object"
                        },
                        "maxItems": 100,
                        "example": [
                            {
                                "option1": "Red",
                                "option2": "Small",
                                "option3": "Round"
                            }
                        ]
                    },
                    "images": {
                        "description": "Product images",
                        "type": "array",
                        "items": {
                            "description": "Image URL",
                            "type": "string",
                            "maxLength": 255,
                            "example": "https://example.com/image.jpg"
                        },
                        "maxItems": 100,
                        "example": [
                            [
                                "https://example.com/image.jpg"
                            ]
                        ]
                    },
                    "description": {
                        "description": "Product description",
                        "type": "string",
                        "example": "Product description"
                    }
                },
                "type": "object"
            },
            "CollectResource": {
                "title": "Collect",
                "description": "Collect model",
                "properties": {
                    "id": {
                        "description": "Collect id",
                        "type": "uuid",
                        "example": "ea9f1b1a-1b1a-11eb-adc1-0242ac120002"
                    },
                    "collection_id": {
                        "description": "Collect collection_id",
                        "type": "uuid",
                        "example": "ea9f1b1a-1b1a-11eb-adc1-0242ac120002"
                    },
                    "product_id": {
                        "description": "Collect product_id",
                        "type": "uuid",
                        "example": "ea9f1b1a-1b1a-11eb-adc1-0242ac120002"
                    },
                    "created_at": {
                        "description": "Collect created at",
                        "type": "string",
                        "example": "2020-11-12 12:00:00"
                    },
                    "updated_at": {
                        "description": "Collect updated at",
                        "type": "string",
                        "example": "2020-11-12 12:00:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Collect"
                }
            },
...
            "ProductImageResource": {
                "title": "ProductImage",
                "description": "ProductImage model",
                "properties": {
                    "id": {
                        "description": "ProductImage id",
                        "type": "uuid",
                        "example": "ea9f1b1a-1b1a-11eb-adc1-0242ac120002"
                    },
                    "product_id": {
                        "description": "ProductImage product_id",
                        "type": "uuid",
                        "example": "ea9f1b1a-1b1a-11eb-adc1-0242ac120002"
                    },
                    "src": {
                        "description": "ProductImage src",
                        "type": "string",
                        "example": "/storage/product_images/1.jpg"
                    },
                    "position": {
                        "description": "ProductImage position",
                        "type": "integer",
                        "example": "1"
                    },
                    "alt": {
                        "description": "ProductImage alt",
                        "type": "string",
                        "example": "ProductImage alt"
                    },
                    "width": {
                        "description": "ProductImage width",
                        "type": "integer",
                        "example": "1"
                    },
                    "height": {
                        "description": "ProductImage height",
                        "type": "integer",
                        "example": "1"
                    },
                    "created_at": {
                        "description": "ProductImage created at",
                        "type": "string",
                        "example": "2020-11-12 12:00:00"
                    },
                    "updated_at": {
                        "description": "ProductImage updated at",
                        "type": "string",
                        "example": "2020-11-12 12:00:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ProductImage"
                }
            },
            "ProductOptionResource": {
                "title": "ProductOption",
                "description": "ProductOption model",
                "properties": {
                    "id": {
                        "description": "ProductOption id",
                        "type": "uuid",
                        "example": "ea9f1b1a-1b1a-11eb-adc1-0242ac120002"
                    },
                    "product_id": {
                        "description": "ProductOption product_id",
                        "type": "uuid",
                        "example": "ea9f1b1a-1b1a-11eb-adc1-0242ac120002"
                    },
                    "name": {
                        "description": "ProductOption name",
                        "type": "string",
                        "example": "ProductOption name"
                    },
                    "values": {
                        "description": "ProductOption values",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "M"
                        },
                        "example": [
                            "M",
                            "L",
                            "XL"
                        ]
                    },
                    "created_at": {
                        "description": "ProductOption created at",
                        "type": "string",
                        "example": "2020-11-12 12:00:00"
                    },
                    "updated_at": {
                        "description": "ProductOption updated at",
                        "type": "string",
                        "example": "2020-11-12 12:00:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ProductOption"
                }
            },
            "ProductResource": {
                "title": "Product",
                "description": "Product model",
                "properties": {
                    "id": {
                        "description": "Product id",
                        "type": "uuid",
                        "example": "ea9f1b1a-1b1a-11eb-adc1-0242ac120002"
                    },
                    "title": {
                        "description": "Product title",
                        "type": "string",
                        "example": "Product title"
                    },
                    "status": {
                        "description": "Product status",
                        "type": "string",
                        "example": "Product status"
                    },
                    "options": {
                        "description": "Product options",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductOptionResource"
                        }
                    },
                    "variants": {
                        "description": "Product variants",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VariantResource"
                        }
                    },
                    "images": {
                        "description": "Product images",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductImageResource"
                        }
                    },
                    "image": {
                        "description": "Product image",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductImageResource"
                        }
                    },
                    "product_type": {
                        "description": "Product product_type",
                        "type": "string",
                        "example": "Product product_type"
                    },
                    "vendor": {
                        "description": "Product vendor",
                        "type": "string",
                        "example": "Product vendor"
                    },
                    "created_at": {
                        "description": "Product created_at",
                        "type": "string",
                        "example": "Product created_at"
                    },
                    "updated_at": {
                        "description": "Product updated_at",
                        "type": "string",
                        "example": "Product updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Product"
                }
            },
            "VariantResource": {
                "title": "Variant",
                "description": "Variant model",
                "properties": {
                    "id": {
                        "description": "Variant id",
                        "type": "uuid",
                        "example": "ea9f1b1a-1b1a-11eb-adc1-0242ac120002"
                    },
                    "product_id": {
                        "description": "Variant product_id",
                        "type": "uuid",
                        "example": "ea9f1b1a-1b1a-11eb-adc1-0242ac120002"
                    },
                    "title": {
                        "description": "Variant title",
                        "type": "string",
                        "example": "Variant title"
                    },
                    "price": {
                        "description": "Variant price",
                        "type": "string",
                        "example": "Variant price"
                    },
                    "gram": {
                        "description": "Variant gram",
                        "type": "string",
                        "example": "Variant gram"
                    },
                    "option1": {
                        "description": "Variant option1",
                        "type": "string",
                        "example": "Variant option1"
                    },
                    "option2": {
                        "description": "Variant option2",
                        "type": "string",
                        "example": "Variant option2"
                    },
                    "option3": {
                        "description": "Variant option3",
                        "type": "string",
                        "example": "Variant option3"
                    },
                    "created_at": {
                        "description": "Variant created at",
                        "type": "string",
                        "example": "2020-11-12 12:00:00"
                    },
                    "updated_at": {
                        "description": "Variant updated at",
                        "type": "string",
                        "example": "2020-11-12 12:00:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Variant"
                }
            }
        }
    }
}